Create lib/entities/doors.lua Move door related methods into doors.lua
